添加目录，文章分页，文章搜索

+++ article/views.py
@@ -4,29 +4,60 @@ from .models import ArticlePost
 import markdown
 from .forms import ArticlePostForm
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.core.paginator import Paginator
+from django.db.models import Q
 
 
 # Create your views here.
 
 
 def article_list(request):
-    articles = ArticlePost.objects.all()
-    context = {'articles': articles}
+    search = request.GET.get('search')
+    order = request.GET.get('order')
+    if search:
+        if order == 'total_views':
+            articles_list = ArticlePost.objects.filter(
+                Q(title__icontains=search) |
+                Q(body__icontains=search)
+            ).order_by('-total_views')
+        else:
+            articles_list = ArticlePost.objects.filter(
+                Q(title__icontains=search) |
+                Q(body__icontains=search)
+            )
+    else:
+        search = ''
+        if order == 'total_views':
+            articles_list = ArticlePost.objects.all().order_by('-total_views')
+        else:
+            articles_list = ArticlePost.objects.all()
+    paginator = Paginator(articles_list, 3)
+    page = request.GET.get('page')
+    articles = paginator.get_page(page)
+
+    # 增加 search 到 context
+    context = {'articles': articles, 'order': order, 'search': search}
+
     return render(request, 'article/list.html', context)
 
 
 def article_detail(request, id):
     article = ArticlePost.objects.get(id=id)
-    article.body = markdown.markdown(article.body,
-                                     extensions=[
-                                         'markdown.extensions.extra',
-                                         'markdown.extensions.codehilite',
-                                     ])
-    context = {'article': article}
+    article.total_views += 1
+    article.save(update_fields=['total_views'])
+    md = markdown.Markdown(extensions=[
+        'markdown.extensions.extra',
+        'markdown.extensions.codehilite',
+        'markdown.extensions.toc',
+    ])
+    article.body = md.convert(article.body)
+    context = {'article': article, 'toc': md.toc}
     return render(request, 'article/detail.html', context)
 
 
 # 写文章额度视图
+@login_required(login_url='userprofile/login/')
 def article_create(request):
     # 判断是否提交数据
     if request.method == "POST":
@@ -39,7 +70,7 @@ def article_create(request):
             # 指定数据库中 id=1 的用户为作者
             # 如果进行过删除数据表的操作， 可能会找不到id=1的用户
             # 此时请重新创建用户，并传入此用户id
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             # 将新文章保存到数据库
             new_article.save()
             # 完成后返回到文章列表
@@ -57,15 +88,19 @@ def article_create(request):
         return render(request, 'article/create.html', context)
 
 
+@login_required(login_url='userprofile/login/')
 def article_safe_delete(request, id):
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
+        if request.user != article.author:
+            return HttpResponse("抱歉，你无权删除这篇文章。")
         article.delete()
         return redirect("article:article_list")
     else:
         return HttpResponse("仅允许post请求")
 
 
+@login_required(login_url='userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -75,6 +110,8 @@ def article_update(request, id):
     """
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     # 判断用户是否为 POST 提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
@@ -93,6 +130,5 @@ def article_update(request, id):
         # 创建表单类实例
         article_post_form = ArticlePostForm()
         # 赋值上下文，将 article文章对象传递进去， 以便提取旧的内容
-        context = { 'article': article, 'article_post_form': article_post_form }
+        context = {'article': article, 'article_post_form': article_post_form}
         return render(request, 'article/update.html', context)
-
